% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamerElements.R
\name{beamerBoxPlot}
\alias{beamerBoxPlot}
\title{Produces box-and-whisker plots of treatments of an experiment (eps-file).}
\usage{
beamerBoxPlot(
  treatments,
  variable,
  title = "",
  outline = TRUE,
  notch = FALSE,
  horizontal = TRUE,
  varwidth = FALSE,
  name = "",
  inpath = ".",
  outpath = "."
)
}
\arguments{
\item{treatments}{List of filenames of treatments.}

\item{variable}{Name of variable.}

\item{title}{Title of Boxplot. Default: "".}

\item{outline}{Boolean. If FALSE, remove outliers. Default: TRUE.}

\item{notch}{Boolean. If TRUE, notches are drawn. Default: FALSE.
Nonoverlapping notches indicate different 
medians of treatments.}

\item{horizontal}{Boolean. Default: TRUE. Plots are horizontal.}

\item{varwidth}{Boolean. Default: FALSE.
If TRUE,  width of the plot indicates sample size.}

\item{name}{Name of experiment. Default: "".}

\item{inpath}{Path to treatment files of experiment.
Default: ".".}

\item{outpath}{Path to report files.}
}
\value{
Name of eps file with output.
}
\description{
Produces box-and-whisker plots of treatments of an experiment (eps-file).
}
\examples{
cat("Example: TODO.\n")

}
\seealso{
Other beamer Elements: 
\code{\link{beamerEndDocument}()},
\code{\link{beamerFigure}()},
\code{\link{beamerMetaTable}()},
\code{\link{beamerProlog}()},
\code{\link{beamerTable}()},
\code{\link{beamerTables}()}
}
\concept{beamer Elements}
