% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamerElements.R
\name{beamerEndDocument}
\alias{beamerEndDocument}
\title{End of document (Latex beamer)}
\usage{
beamerEndDocument()
}
\value{
A text string with latex markup.
}
\description{
End of document (Latex beamer)
}
\examples{
a<-beamerEndDocument()
print(a)
cat(a)
}
\seealso{
Other beamer Elements: 
\code{\link{beamerBoxPlot}()},
\code{\link{beamerFigure}()},
\code{\link{beamerMetaTable}()},
\code{\link{beamerProlog}()},
\code{\link{beamerTable}()},
\code{\link{beamerTables}()}
}
\concept{beamer Elements}
