% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamerElements.R
\name{beamerMetaTable}
\alias{beamerMetaTable}
\title{The meta table environment for a latex document.}
\usage{
beamerMetaTable(name, fnName, caption = "No caption defined", label = "")
}
\arguments{
\item{name}{Name of the experiment.}

\item{fnName}{File name of the exported table element.}

\item{caption}{Caption of the table.}

\item{label}{Label of the table.}
}
\value{
A text string with latex markup.
}
\description{
The meta table environment for a latex document.
}
\examples{
a<-beamerMetaTable(name="Ex A", fnName="test.tex", caption="test", label="ltest")
print(a)
cat(a)

}
\seealso{
Other beamer Elements: 
\code{\link{beamerBoxPlot}()},
\code{\link{beamerEndDocument}()},
\code{\link{beamerFigure}()},
\code{\link{beamerProlog}()},
\code{\link{beamerTable}()},
\code{\link{beamerTables}()}
}
\concept{beamer Elements}
