% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamerElements.R
\name{beamerSection}
\alias{beamerSection}
\title{Sections of a latex beamer document.}
\usage{
beamerSection(name, level = "section", clearpage = FALSE, miniframe = TRUE)
}
\arguments{
\item{name}{Name of the document element.}

\item{level}{Latex markup of the document element.
Default: "section". Latex supports:
"section", "subsection", "subsubsection".}

\item{clearpage}{Boolean. Default: TRUE. 
Clearpage before start of document element.
Forces output of all floating elements of 
the previous document element.}

\item{miniframe}{Boolean. Default: TRUE.
If FALSE, miniframes are turned off.}
}
\value{
A text string with latex markup.
}
\description{
Sections of a latex beamer document.
}
\examples{
beamerSection(name="Design", level="section", clearpage=FALSE)
a<-beamerSection(name="Design", level="section", clearpage=TRUE)
print(a)
cat(a)
}
\seealso{
Other Latex Elements: 
\code{\link{latexBoxPlot}()},
\code{\link{latexEndDocument}()},
\code{\link{latexFigure}()},
\code{\link{latexMetaTable}()},
\code{\link{latexProlog}()},
\code{\link{latexSection}()},
\code{\link{latexTable}()},
\code{\link{latexTables}()},
\code{\link{latexText}()}
}
\concept{Latex Elements}
