% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamerElements.R
\name{beamerTable}
\alias{beamerTable}
\title{The table environment for a latex document.}
\usage{
beamerTable(
  df,
  caption = "",
  label = "",
  align = "",
  center = "center",
  sanitize = FALSE,
  tableLength = 20
)
}
\arguments{
\item{df}{Dataframe or matrix.}

\item{caption}{Caption of the table. Default: "".}

\item{label}{Label of the table.   Default: "".}

\item{align}{Alignment of columns. Default: "".}

\item{center}{Center table. Default: "center".}

\item{sanitize}{Boolean. Default: FALSE.}

\item{tableLength}{Integer. Default: 25.}
}
\value{
A list of text strings with latex markup.
        Each list element is a table.
}
\description{
The table environment for a latex document.
}
\examples{
df<-data.frame(Name=c("Alice", "Bob"), Age=c(25, 27))
a<-beamerTable(df, caption="Alice and Bob", label="AB")
print(a)

}
\seealso{
Other beamer Elements: 
\code{\link{beamerBoxPlot}()},
\code{\link{beamerEndDocument}()},
\code{\link{beamerFigure}()},
\code{\link{beamerMetaTable}()},
\code{\link{beamerProlog}()},
\code{\link{beamerTables}()}
}
\concept{beamer Elements}
