% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentElements.R
\name{beamerTableLength}
\alias{beamerTableLength}
\title{Beamer table length.}
\usage{
beamerTableLength()
}
\value{
Integer. Number of lines in beamer table.
}
\description{
Beamer table length.
}
\examples{
beamerTableLength()

}
\seealso{
Other Internal Configuration: 
\code{\link{latexTableLength}()}
}
\concept{Internal Configuration}
