% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamerElements.R
\name{beamerText}
\alias{beamerText}
\title{Text.}
\usage{
beamerText(..., header = "Comment", block = TRUE)
}
\arguments{
\item{...}{A list of text strings.}

\item{header}{Name of the block.}

\item{block}{Default: TRUE.}
}
\value{
A text string.
}
\description{
Text.
}
\examples{
a<-beamerText("Here", "we", "are")
print(a)
cat(a)
}
\seealso{
Other Console Elements: 
\code{\link{consoleBoxPlot}()},
\code{\link{consoleEndDocument}()},
\code{\link{consoleFigure}()},
\code{\link{consoleProlog}()},
\code{\link{consoleSection}()},
\code{\link{consoleTable}()},
\code{\link{consoleText}()},
\code{\link{pause}()}
}
\concept{Console Elements}
