% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kSymmetryWrapper.R
\name{booleanGrammarK}
\alias{booleanGrammarK}
\title{booleanGrammarK}
\usage{
booleanGrammarK()
}
\description{
booleanGrammarK
}
\seealso{
Other xegaBNF: 
\code{\link{variableNamesLHS}()},
\code{\link{writeBNF}()}
}
\concept{xegaBNF}
