% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentElements.R
\name{buildLatexMain}
\alias{buildLatexMain}
\title{Build latex main program.}
\usage{
buildLatexMain(name, path = ".")
}
\arguments{
\item{name}{Name of experiment.}

\item{path}{Path of latex report files.}
}
\value{
List of imported files. (invisible).
}
\description{
\code{buildLatexMain()}
             reads all latex files starting with "main" from the 
             directory specified by path, catenates them and writes
             a new latex file "<main>main.tex" to the same directory.
}
\concept{Report of Experiment (Internal)}
