% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consoleElements.R
\name{consoleEndDocument}
\alias{consoleEndDocument}
\title{End of document (LaTeX)}
\usage{
consoleEndDocument()
}
\value{
A text string (invisible).
}
\description{
End of document (LaTeX)
}
\examples{
a<-consoleEndDocument()
print(a)
cat(a)
}
\seealso{
Other Console Elements: 
\code{\link{beamerText}()},
\code{\link{consoleBoxPlot}()},
\code{\link{consoleFigure}()},
\code{\link{consoleProlog}()},
\code{\link{consoleSection}()},
\code{\link{consoleTable}()},
\code{\link{consoleText}()},
\code{\link{pause}()}
}
\concept{Console Elements}
