% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consoleElements.R
\name{consoleFigure}
\alias{consoleFigure}
\title{The figure output on a console.}
\usage{
consoleFigure(
  name = "",
  fnName = "",
  caption = "No caption defined",
  label = ""
)
}
\arguments{
\item{name}{Name of the experiment.}

\item{fnName}{File name of the eps file of the graphic.}

\item{caption}{Caption of the figure.}

\item{label}{Label of the figure.}
}
\value{
A text string (invisible).
}
\description{
The figure output on a console.
}
\examples{
consoleFigure(name="Ex A", fnName="test.eps", caption="test", label="ltest")

}
\seealso{
Other Console Elements: 
\code{\link{beamerText}()},
\code{\link{consoleBoxPlot}()},
\code{\link{consoleEndDocument}()},
\code{\link{consoleProlog}()},
\code{\link{consoleSection}()},
\code{\link{consoleTable}()},
\code{\link{consoleText}()},
\code{\link{pause}()}
}
\concept{Console Elements}
