% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consoleElements.R
\name{consoleProlog}
\alias{consoleProlog}
\title{Prolog of experiment.}
\usage{
consoleProlog(
  name = "Missing",
  title = "Undefined",
  author = "Unknown",
  purpose = "Unknown"
)
}
\arguments{
\item{name}{Name of experiment. Default: "Missing".}

\item{title}{Title of the experiment. Default: "Undefined".}

\item{author}{Author of experimental report. Default: "Unknown".}

\item{purpose}{Purpose of experiment. Default: "Unknown".}
}
\value{
A text string (invisible).
}
\description{
Shows name, title, author, and purpose of experiment.
}
\examples{
consoleProlog(name="Experiment A.", 
    title="First test", author="U.N. Owen", 
    purpose=paste0("The experiment shows that algorithm A ", 
                   "is faster than algorithm B.\n"))

}
\seealso{
Other Console Elements: 
\code{\link{beamerText}()},
\code{\link{consoleBoxPlot}()},
\code{\link{consoleEndDocument}()},
\code{\link{consoleFigure}()},
\code{\link{consoleSection}()},
\code{\link{consoleTable}()},
\code{\link{consoleText}()},
\code{\link{pause}()}
}
\concept{Console Elements}
