% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consoleElements.R
\name{consoleTable}
\alias{consoleTable}
\title{The table environment for the console.}
\usage{
consoleTable(df, caption = "", label = "", align = "", center = "center")
}
\arguments{
\item{df}{Dataframe or matrix.}

\item{caption}{Caption of the table. Default: "".}

\item{label}{Label of the table.   Default: "".}

\item{align}{Alignment of columns. Default: "".}

\item{center}{Center table. Default: "center".}
}
\value{
A dataframe (invisible).
}
\description{
The table environment for the console.
}
\examples{
df<-data.frame(Name=c("Alice", "Bob"), Age=c(25, 27))
consoleTable(df, caption="Alice and Bob", label="AB")

}
\seealso{
Other Console Elements: 
\code{\link{beamerText}()},
\code{\link{consoleBoxPlot}()},
\code{\link{consoleEndDocument}()},
\code{\link{consoleFigure}()},
\code{\link{consoleProlog}()},
\code{\link{consoleSection}()},
\code{\link{consoleText}()},
\code{\link{pause}()}
}
\concept{Console Elements}
