% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consoleElements.R
\name{consoleText}
\alias{consoleText}
\title{Text.}
\usage{
consoleText(..., header = "")
}
\arguments{
\item{...}{A list of text strings.}

\item{header}{Name of the block.}
}
\value{
A text string (invisible).
}
\description{
Text.
}
\examples{
a<-consoleText("here", "we", "are")
print(a)
cat(a)
}
\seealso{
Other Console Elements: 
\code{\link{beamerText}()},
\code{\link{consoleBoxPlot}()},
\code{\link{consoleEndDocument}()},
\code{\link{consoleFigure}()},
\code{\link{consoleProlog}()},
\code{\link{consoleSection}()},
\code{\link{consoleTable}()},
\code{\link{pause}()}
}
\concept{Console Elements}
