% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentDesign.R
\name{designOfAllTreatments}
\alias{designOfAllTreatments}
\title{Generate a dataframe with all parameters of all treatments of an experiment.}
\usage{
designOfAllTreatments(treatments, inpath = ".")
}
\arguments{
\item{treatments}{List of treatment names of experiments.}

\item{inpath}{Path for treatment files. Default: ".".}
}
\value{
A dataframe with all parameters of all treatments.
}
\description{
Generate a dataframe with all parameters of all treatments of an experiment.
}
\examples{
cat("TBD!\n")

}
\seealso{
Other Design of Experiment: 
\code{\link{designOfExperiment}()},
\code{\link{designOfTreatment}()},
\code{\link{designSplitParameter}()},
\code{\link{getTreatment}()},
\code{\link{pl2df}()}
}
\concept{Design of Experiment}
