% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentDesign.R
\name{designOfExperiment}
\alias{designOfExperiment}
\title{Design of Experiment.}
\usage{
designOfExperiment(treatments, inpath = ".")
}
\arguments{
\item{treatments}{List of treatment names of experiments.}

\item{inpath}{Path for treatment files. Default: ".".}
}
\value{
A list with two dataframes
        \itemize{
        \item \code{$common} Dataframe of parameters common to all treatments of an experiment.
        \item \code{$different} Dataframe of parameters different in the treatments of an experiment.
        }
}
\description{
Design of Experiment.
}
\examples{
cat("TBD!\n")

}
\seealso{
Other Design of Experiment: 
\code{\link{designOfAllTreatments}()},
\code{\link{designOfTreatment}()},
\code{\link{designSplitParameter}()},
\code{\link{getTreatment}()},
\code{\link{pl2df}()}
}
\concept{Design of Experiment}
