% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentDesign.R
\name{designOfTreatment}
\alias{designOfTreatment}
\title{Return a dataframe with all parameters of a treatment.}
\usage{
designOfTreatment(treatmentdf)
}
\arguments{
\item{treatmentdf}{A dataframe with the results of a treatment.}
}
\value{
A dataframe with all parameters of a treatment.
}
\description{
Return a dataframe with all parameters of a treatment.
}
\examples{
cat("TBD!\n")

}
\seealso{
Other Design of Experiment: 
\code{\link{designOfAllTreatments}()},
\code{\link{designOfExperiment}()},
\code{\link{designSplitParameter}()},
\code{\link{getTreatment}()},
\code{\link{pl2df}()}
}
\concept{Design of Experiment}
