% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentDesign.R
\name{designSplitParameter}
\alias{designSplitParameter}
\title{Split dataframe in common and different parameters.}
\usage{
designSplitParameter(DF)
}
\arguments{
\item{DF}{Dataframe with all parameters of all treatments of an experiment.}
}
\value{
A list with two dataframes
        \itemize{
        \item \code{$common} Dataframe of parameters common to all treatments of an experiment.
        \item \code{$different} Dataframe of parameters different in the treatments of an experiment.
        }
}
\description{
Split dataframe in common and different parameters.
}
\examples{
cat("TBD!\n")

}
\seealso{
Other Design of Experiment: 
\code{\link{designOfAllTreatments}()},
\code{\link{designOfExperiment}()},
\code{\link{designOfTreatment}()},
\code{\link{getTreatment}()},
\code{\link{pl2df}()}
}
\concept{Design of Experiment}
