% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envKsymmetry.R
\docType{data}
\name{env5symmetry}
\alias{env5symmetry}
\title{A problem environment for the 5-symmetry function.}
\format{
An object of class \code{list} of length 7.
}
\usage{
env5symmetry
}
\description{
A problem environment for the 5-symmetry function.
}
\examples{
t3<-"AND(OR(AND(D2,D4),AND(NOT(D2),NOT(D4)))," 
t3<-paste0(t3,"OR(AND(D1, D5), AND(NOT(D1), NOT(D5))))")
env5symmetry$f(t1)
env5symmetry$f(t3)

}
\seealso{
Other Problem Environment: 
\code{\link{env5symmetryNN}}
}
\concept{Problem Environment}
\keyword{datasets}
