% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envKsymmetryNN.R
\docType{data}
\name{env5symmetryNN}
\alias{env5symmetryNN}
\title{A problem environment for the 5-symmetry function.}
\format{
An object of class \code{list} of length 16.
}
\usage{
env5symmetryNN
}
\description{
A problem environment for the 5-symmetry function.
}
\examples{
env5symmetryNN$f(rndParms(env5symmetryNN$nNNparms()))

}
\seealso{
Other Problem Environment: 
\code{\link{env5symmetry}}
}
\concept{Problem Environment}
\keyword{datasets}
