% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envRealFunctions-package.R
\name{envRealFunctions}
\alias{envRealFunctions}
\title{Package Problem Environments for Real Functions}
\description{
Problem environments with 
real test functions for simple genetic algorithms
}
\details{
The \code{envRealFunctions} package implements
for example the functions F1-F5 of Kenneth De Jong's Thesis.

Problem environments for the package \code{xega} must be implemented as 
lists of functions or function factories which return with at least 
the following functions:
\itemize{
\item \code{<factory name>$name()} 
      is a constant function which returns 
      the name of the environment.
\item \code{<factory name>$bitlength()} is a constant function which returns
      a vector which returns for each parameter the number of bits 
      used for coding the parameter value.
\item \code{<factory name>$genelength()} is a constant function which 
      returns the number of bits of a gene. 
\item \code{<factory name>$lb()} is a constant function which returns
      the vector of lower bounds of the parameters.
\item \code{<factory name>$ub()} is a constant function which returns
      the vector of upper bounds of the parameters.
\item \code{<factory name>$f(parm, gene, lF)} specifies the function  
      with the parameter vector \code{parm} which should be 
      optimized. The variables \code{gene} and \code{lF} 
      extend the interface between genetic algorithm and problem 
      environment. Both variables are not used in this package.
}

The constant functions 
\code{<factory name>$bitlength()},
\code{<factory name>$genelength()},
\code{<factory name>$lb()},
\code{<factory name>$ub()}
specify the precision of the parameters as well as
the hypercube in which the optimal solution(s) are searched.
These four functions provide the information needed for decoding
binary genes. For real coded genes, 
\code{length(<factory name>$bitlength())} indicates the number of 
parameters.

Three additional functions can be provided:
\itemize{
\item \code{<factory name>$terminate(solution)} TRUE, if the solution 
           meets an early termination condition.
\item \code{<factory name>$describe()} prints a description of the 
      function (known mathematical properties and a reference to the
      origin of the function).
\item \code{<factory name>$solution()} specifies 
      the global minimum/maximum solution value and
      the list of minimal/maximal points.
      The function \code{penvValidate}(penv) validates the correct
      implementation of the function \code{f} of the environment.
}

At the moment, the package contains: 
  \enumerate{
   \item A set of 2-D parabola functions in DelayedP.R for testing:
         With time delays, with random errors, and early termination.
   \item De Jong's functions (F1-F5).
   \item Multimodal 2-D sine waves (AGSF7d1, F6, F6inv, and F7).
   \item Functions A1-A30 from Lavezzi et al. (2022).
   \item A wrapper for the single-objective optimization test functions
         of the smoof package of Bossek (2017).
   }
}
\section{Copyright}{
 (c) 2023 Andreas Geyer-Schulz
}

\section{License}{
 MIT
}

\section{URL}{
 TBD
}

\section{Installation}{
 TBD
}

\references{
De Jong, Kenneth A. (1975):
 \emph{An Analysis of the Behavior of a Class of Genetic Adaptive Systems}
 University of Michigan, Ann Arbor.
 URL: https://cs.gmu.edu/~eclab/kdj_thesis.html 

Lavezzi, Giovanni and Guye, Kidus and Ciarcia, Marco (2022):
Nonlinear Programming Solvers for Unconstrained and Constrained
Optimization Problems: a Benchmark Analysis. 
arXiv, arXiv:2204.05297v1 (11 April 2022).

Bossek, Jakob (2017)
         smoof: Single- and Multi-Objective Optimization Test Functions.
          The R Journal, 9(1), 103-113.
}
\author{
Andreas Geyer-Schulz
}
\concept{problem environments}
