% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNnFF.R
\name{envXORNN2421}
\alias{envXORNN2421}
\title{Generate a problem environment for the XOR problem.
Topology: c(2, 4,  2, 1)}
\usage{
envXORNN2421()
}
\description{
Generate a problem environment for the XOR problem.
Topology: c(2, 4,  2, 1)
}
\examples{
require(xega)
cat("\n Topology: c(2, 4, 2, 1). Activation: ReLU \n")
p3<-envXORNN2421()
t3<-xegaRun(penv=p3, max=FALSE, algorithm="sgde", popsize=30, generations=100, 
   mutation="MutateGeneDE", scalefactor="Uniform", crossover="UCrossGene", 
   genemap="Identity", replication="DE", 
   selection="UniformP", mateselection="UniformP", accept="Best")
printNNweights(t3$solution$phenotype, p3$topology())
as<-NN(t3$solution$phenotype, p3$topology(), data=xorData)
cat("Fitness:", t3$solution$fitness, "Above 0.5? \n")
r3<-data.frame(as, (as>0.5), xorData[,3])
colnames(r3)<-c("Activation", "Predicted", "xor")
print(r3)

}
