% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentReport.R
\name{experimentAnalysisOfTreatment}
\alias{experimentAnalysisOfTreatment}
\title{Analysis of a treatment.}
\usage{
experimentAnalysisOfTreatment(
  name = "",
  treatmentname = "",
  miniframe = TRUE,
  inpath = ".",
  outpath = ".",
  type = "console"
)
}
\arguments{
\item{name}{Name of Experiment.}

\item{treatmentname}{Name of rds-file of treatment}

\item{miniframe}{Boolean. Default: TRUE.
If false, turn off miniframes.}

\item{inpath}{Path to treatment rds-files of experiment.
Default: ".".}

\item{outpath}{Path to report of experiment. Default: "."}

\item{type}{Output to console (type="console") 
or to latex files (type="latex")?  
Default: "console".}
}
\value{
0 (invisible).
}
\description{
Analysis of a treatment.
}
\examples{
cat("TBD! \n")

}
\seealso{
Other Report of Experiment: 
\code{\link{experimentAllXegaParameters}()},
\code{\link{experimentAppendix}()},
\code{\link{experimentBoxPlot}()},
\code{\link{experimentDesign}()},
\code{\link{experimentEffectReport}()},
\code{\link{experimentEnd}()},
\code{\link{experimentGrammarTable}()},
\code{\link{experimentHypothesisTest}()},
\code{\link{experimentIsNormal}()},
\code{\link{experimentMeanMatrix}()},
\code{\link{experimentNNTables}()},
\code{\link{experimentPlotPopStats}()},
\code{\link{experimentSection}()},
\code{\link{experimentSolutionGenotype}()},
\code{\link{experimentSolutionTable}()},
\code{\link{experimentStart}()},
\code{\link{experimentStatistic}()},
\code{\link{experimentText}()},
\code{\link{experimentTreatmentParameters}()},
\code{\link{experimentTreatments}()},
\code{\link{experimentXegaParameters}()}
}
\concept{Report of Experiment}
