% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentAppendix.R
\name{experimentAppendix}
\alias{experimentAppendix}
\title{Appendix of an experimental report.}
\usage{
experimentAppendix(
  treatments,
  name = "",
  miniframe = TRUE,
  inpath = ".",
  outpath = ",",
  type = "console"
)
}
\arguments{
\item{treatments}{List of treatment file names.}

\item{name}{Name of experiment. Default: "".}

\item{miniframe}{Boolean. Default: TRUE.
If false, turn off miniframes in treatment subsection.}

\item{inpath}{Path to treatment rds-files of experiment.
Default: ".".}

\item{outpath}{Path to report of experiment. Default: "."}

\item{type}{Output to console (type="console") 
or to latex files (type="latex" | "beamer")?  
Default: "console".}
}
\value{
0 (invisible).
}
\description{
Three parts:
  \enumerate{
  \item Summary of descriptive statistics of the distribution 
        of the performance
        variables of all treatments for n trials.
  \item Analysis of each treatment. 
  \item All xega parameters (and their setting for one treatment).
  }
}
\examples{
cat("TBD!\n")

}
\seealso{
Other Report of Experiment: 
\code{\link{experimentAllXegaParameters}()},
\code{\link{experimentAnalysisOfTreatment}()},
\code{\link{experimentBoxPlot}()},
\code{\link{experimentDesign}()},
\code{\link{experimentEffectReport}()},
\code{\link{experimentEnd}()},
\code{\link{experimentGrammarTable}()},
\code{\link{experimentHypothesisTest}()},
\code{\link{experimentIsNormal}()},
\code{\link{experimentMeanMatrix}()},
\code{\link{experimentNNTables}()},
\code{\link{experimentPlotPopStats}()},
\code{\link{experimentSection}()},
\code{\link{experimentSolutionGenotype}()},
\code{\link{experimentSolutionTable}()},
\code{\link{experimentStart}()},
\code{\link{experimentStatistic}()},
\code{\link{experimentText}()},
\code{\link{experimentTreatmentParameters}()},
\code{\link{experimentTreatments}()},
\code{\link{experimentXegaParameters}()}
}
\concept{Report of Experiment}
