% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentElements.R
\name{experimentBoxPlot}
\alias{experimentBoxPlot}
\title{Produces box-and-whisker plots of treatments of an experiment.}
\usage{
experimentBoxPlot(
  treatments,
  variable,
  title = "",
  outline = TRUE,
  notch = FALSE,
  horizontal = TRUE,
  varwidth = FALSE,
  name = "",
  inpath = ".",
  outpath = ".",
  type = "console"
)
}
\arguments{
\item{treatments}{List of filenames of treatments.}

\item{variable}{Name of variable.}

\item{title}{Title of Boxplot. Default: "".}

\item{outline}{Boolean. If FALSE, remove outliers. Default: TRUE.}

\item{notch}{Boolean. If TRUE, notches are drawn. Default: FALSE.
Nonoverlapping notches indicate different 
medians of treatments.}

\item{horizontal}{Boolean. Default: TRUE. Plots are horizontal.}

\item{varwidth}{Boolean. Default: FALSE.
If TRUE,  width of the plot indicates sample size.}

\item{name}{Name of experiment. Default: "".}

\item{inpath}{Path to treatment files of experiment.
Default: ".".}

\item{outpath}{Path to report files.}

\item{type}{Output to console (type="console") 
or to latex files (type="latex" | "beamer")?  
Default: "console".}
}
\value{
0 (invisible).
}
\description{
See the documentation of base::boxplot.
}
\examples{
cat("Example: TODO.\n")

}
\seealso{
Other Report of Experiment: 
\code{\link{experimentAllXegaParameters}()},
\code{\link{experimentAnalysisOfTreatment}()},
\code{\link{experimentAppendix}()},
\code{\link{experimentDesign}()},
\code{\link{experimentEffectReport}()},
\code{\link{experimentEnd}()},
\code{\link{experimentGrammarTable}()},
\code{\link{experimentHypothesisTest}()},
\code{\link{experimentIsNormal}()},
\code{\link{experimentMeanMatrix}()},
\code{\link{experimentNNTables}()},
\code{\link{experimentPlotPopStats}()},
\code{\link{experimentSection}()},
\code{\link{experimentSolutionGenotype}()},
\code{\link{experimentSolutionTable}()},
\code{\link{experimentStart}()},
\code{\link{experimentStatistic}()},
\code{\link{experimentText}()},
\code{\link{experimentTreatmentParameters}()},
\code{\link{experimentTreatments}()},
\code{\link{experimentXegaParameters}()}
}
\concept{Report of Experiment}
