% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentElements.R
\name{experimentDesign}
\alias{experimentDesign}
\title{Produces tables of common and different parameters of treatments.}
\usage{
experimentDesign(
  treatments,
  inpath = ".",
  outpath = ".",
  caption = "",
  name = "",
  type = "console"
)
}
\arguments{
\item{treatments}{List of treatment names of experiment.}

\item{inpath}{Path to treatment rds-files of experiment.
Default: ".".}

\item{outpath}{Path to report of experiment. Default: "."}

\item{caption}{Caption of table. Default: "".
Default: ".".}

\item{name}{Name of experiment. Default: "".}

\item{type}{Output to console (type="console") 
or to latex files (type="latex" | "beamer")?  
Default: "console".}
}
\value{
0 (invisible).
}
\description{
The table of different parameters of treatments
             and the table of common parameters of treatments
             are traditionally called the design of the experiment.
             The common parameters should make the experiment replicable,
             The behaviorally relevant different parameters provide
             the basis of testing hypothesis about the effects 
             caused by these parameters.
}
\examples{
cat("TBD! \n")

}
\seealso{
Other Report of Experiment: 
\code{\link{experimentAllXegaParameters}()},
\code{\link{experimentAnalysisOfTreatment}()},
\code{\link{experimentAppendix}()},
\code{\link{experimentBoxPlot}()},
\code{\link{experimentEffectReport}()},
\code{\link{experimentEnd}()},
\code{\link{experimentGrammarTable}()},
\code{\link{experimentHypothesisTest}()},
\code{\link{experimentIsNormal}()},
\code{\link{experimentMeanMatrix}()},
\code{\link{experimentNNTables}()},
\code{\link{experimentPlotPopStats}()},
\code{\link{experimentSection}()},
\code{\link{experimentSolutionGenotype}()},
\code{\link{experimentSolutionTable}()},
\code{\link{experimentStart}()},
\code{\link{experimentStatistic}()},
\code{\link{experimentText}()},
\code{\link{experimentTreatmentParameters}()},
\code{\link{experimentTreatments}()},
\code{\link{experimentXegaParameters}()}
}
\concept{Report of Experiment}
