% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectReport.R
\name{experimentEffectReport}
\alias{experimentEffectReport}
\title{Report the effects of two different treatments-}
\usage{
experimentEffectReport(
  name = "",
  treatmentname1 = "",
  treatmentname2 = "",
  variable = "",
  test = "wilcox.test",
  mu = 0,
  alternative = "two.sided",
  alpha = 0.05,
  coef = 1.5,
  silent = TRUE,
  inpath = ".",
  outpath = ".",
  type = "console"
)
}
\arguments{
\item{name}{Name of Experiment.}

\item{treatmentname1}{Name of rds-file of treatment 1 (data vector x).}

\item{treatmentname2}{Name of rds-file of treatment 2 (data vector y).}

\item{variable}{Name of variable tested.}

\item{test}{Name of test. Default: "wilcox.test".
Available: "wilcox.test" or "t.test". 
If both data vectors are normal, 
"t.test" is used.}

\item{mu}{Mean of difference of variables of treatments.
Default: 0.}

\item{alternative}{Alternative hypothesis. Default: "two.sided"
\itemize{
\item "two.sided": \code{0==(mean(x) - mean(y))}.
\item "less": \code{(mean(x) - mean(y))} smaller 0.
\item "greate": \code{(mean(x) - mean(y))} greater 0.
}}

\item{alpha}{Level of significance. Default: 0.05.}

\item{coef}{Scaling of window for removing outliers.
Default: 1.5 (as in boxplot.stats). 
0 means all data are retained.
The windox size is given by the length of the box
plus/minus coef times the length of the box.}

\item{silent}{If FALSE, supress some output.}

\item{inpath}{Path to treatment rds-files of experiment.
Default: ".".}

\item{outpath}{Path to report of experiment. Default: "."}

\item{type}{Output to console (type="console") 
or to latex files (type="latex")?  
Default: "console".}
}
\value{
0 (invisible).
}
\description{
Report the effects of two different treatments-
}
\examples{
cat("TBD\n")

}
\seealso{
Other Report of Experiment: 
\code{\link{experimentAllXegaParameters}()},
\code{\link{experimentAnalysisOfTreatment}()},
\code{\link{experimentAppendix}()},
\code{\link{experimentBoxPlot}()},
\code{\link{experimentDesign}()},
\code{\link{experimentEnd}()},
\code{\link{experimentGrammarTable}()},
\code{\link{experimentHypothesisTest}()},
\code{\link{experimentIsNormal}()},
\code{\link{experimentMeanMatrix}()},
\code{\link{experimentNNTables}()},
\code{\link{experimentPlotPopStats}()},
\code{\link{experimentSection}()},
\code{\link{experimentSolutionGenotype}()},
\code{\link{experimentSolutionTable}()},
\code{\link{experimentStart}()},
\code{\link{experimentStatistic}()},
\code{\link{experimentText}()},
\code{\link{experimentTreatmentParameters}()},
\code{\link{experimentTreatments}()},
\code{\link{experimentXegaParameters}()}
}
\concept{Report of Experiment}
