% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentElements.R
\name{experimentIsNormal}
\alias{experimentIsNormal}
\title{Testing for normality of a data vector.}
\usage{
experimentIsNormal(
  treatmentname = "",
  name = "",
  variable = "",
  alpha = 0.05,
  coef = 1.5,
  silent = TRUE,
  inpath = ".",
  outpath = ".",
  type = "console"
)
}
\arguments{
\item{treatmentname}{Name of rds-file of treatment.}

\item{name}{Name of experiment. Default: "".}

\item{variable}{Name of variable. Default: "".}

\item{alpha}{Level of significance. Default: 0.05.}

\item{coef}{Scaling of window for removing outliers.
Default: 1.5 (as in boxplot.stats). 
0 means all data are retained.
The windox size is given by the length of the box
plus/minus coef times the length of the box.}

\item{silent}{If TRUE, no output.}

\item{inpath}{Path to treatment rds-files of experiment.
Default: ".".}

\item{outpath}{Path to report of experiment. Default: "."}

\item{type}{Output to console (type="console") 
or to latex files (type="latex" | "beamer")?  
Default: "console".}
}
\value{
Boolean (invisible).
}
\description{
A wrapper for the Shapiro-Wilk test. 
             See \code{stat::shapiro.test}.
}
\examples{
cat("TBD!\n")

}
\seealso{
Other Report of Experiment: 
\code{\link{experimentAllXegaParameters}()},
\code{\link{experimentAnalysisOfTreatment}()},
\code{\link{experimentAppendix}()},
\code{\link{experimentBoxPlot}()},
\code{\link{experimentDesign}()},
\code{\link{experimentEffectReport}()},
\code{\link{experimentEnd}()},
\code{\link{experimentGrammarTable}()},
\code{\link{experimentHypothesisTest}()},
\code{\link{experimentMeanMatrix}()},
\code{\link{experimentNNTables}()},
\code{\link{experimentPlotPopStats}()},
\code{\link{experimentSection}()},
\code{\link{experimentSolutionGenotype}()},
\code{\link{experimentSolutionTable}()},
\code{\link{experimentStart}()},
\code{\link{experimentStatistic}()},
\code{\link{experimentText}()},
\code{\link{experimentTreatmentParameters}()},
\code{\link{experimentTreatments}()},
\code{\link{experimentXegaParameters}()}
}
\concept{Report of Experiment}
