% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentElements.R
\name{experimentMeanMatrix}
\alias{experimentMeanMatrix}
\title{Produces matrix of mean vectors  of statistics of experiment.}
\usage{
experimentMeanMatrix(
  treatments,
  tpvec = "",
  variable = NULL,
  statistic = "mean",
  inpath = ".",
  outpath = ".",
  caption = "",
  name = "",
  type = "console"
)
}
\arguments{
\item{treatments}{List of treatment names of experiment.}

\item{tpvec}{Vector of treatment patterns.}

\item{variable}{Name of variable.  Default: NULL.}

\item{statistic}{Name of statistic. Default: "mean".}

\item{inpath}{Path to treatment rds-files of experiment.
Default: ".".}

\item{outpath}{Path to report of experiment. Default: "."}

\item{caption}{Caption of table. Default: "".
Default: ".".}

\item{name}{Name of experiment. Default: "".}

\item{type}{Output to console (type="console") 
or to latex files (type="latex" | "beamer")?  
Default: "console".}
}
\value{
0 (invisible).
}
\description{
Specify either treatment name or variable name or 
             none. Produces tables with statistical result 
             of experiment.
}
\examples{
cat("TBD! \n")

}
\seealso{
Other Report of Experiment: 
\code{\link{experimentAllXegaParameters}()},
\code{\link{experimentAnalysisOfTreatment}()},
\code{\link{experimentAppendix}()},
\code{\link{experimentBoxPlot}()},
\code{\link{experimentDesign}()},
\code{\link{experimentEffectReport}()},
\code{\link{experimentEnd}()},
\code{\link{experimentGrammarTable}()},
\code{\link{experimentHypothesisTest}()},
\code{\link{experimentIsNormal}()},
\code{\link{experimentNNTables}()},
\code{\link{experimentPlotPopStats}()},
\code{\link{experimentSection}()},
\code{\link{experimentSolutionGenotype}()},
\code{\link{experimentSolutionTable}()},
\code{\link{experimentStart}()},
\code{\link{experimentStatistic}()},
\code{\link{experimentText}()},
\code{\link{experimentTreatmentParameters}()},
\code{\link{experimentTreatments}()},
\code{\link{experimentXegaParameters}()}
}
\concept{Report of Experiment}
