% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentElements.R
\name{experimentPlotPopStats}
\alias{experimentPlotPopStats}
\title{Produces a plot of population statistics of \code{xegaRun}.}
\usage{
experimentPlotPopStats(
  treatmentname,
  name = "",
  inpath = ".",
  outpath = ".",
  type = "console"
)
}
\arguments{
\item{treatmentname}{Name of rds-file of treatment.}

\item{name}{Name of experiment. Default: "".}

\item{inpath}{Path to treatment files of experiment.
Default: ".".}

\item{outpath}{Path to report files.}

\item{type}{Output to console (type="console") 
or to latex files (type="latex" | "beamer")?  
Default: "console".}
}
\value{
0 (invisible).
}
\description{
See the documentation xegaPlotPopStats.
}
\examples{
cat("Example: TODO.\n")

}
\seealso{
Other Report of Experiment: 
\code{\link{experimentAllXegaParameters}()},
\code{\link{experimentAnalysisOfTreatment}()},
\code{\link{experimentAppendix}()},
\code{\link{experimentBoxPlot}()},
\code{\link{experimentDesign}()},
\code{\link{experimentEffectReport}()},
\code{\link{experimentEnd}()},
\code{\link{experimentGrammarTable}()},
\code{\link{experimentHypothesisTest}()},
\code{\link{experimentIsNormal}()},
\code{\link{experimentMeanMatrix}()},
\code{\link{experimentNNTables}()},
\code{\link{experimentSection}()},
\code{\link{experimentSolutionGenotype}()},
\code{\link{experimentSolutionTable}()},
\code{\link{experimentStart}()},
\code{\link{experimentStatistic}()},
\code{\link{experimentText}()},
\code{\link{experimentTreatmentParameters}()},
\code{\link{experimentTreatments}()},
\code{\link{experimentXegaParameters}()}
}
\concept{Report of Experiment}
