% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentElements.R
\name{experimentSection}
\alias{experimentSection}
\title{Sections of an experiment}
\usage{
experimentSection(
  name = "Missing experiment",
  secname = "Missing section",
  level = "section",
  clearpage = TRUE,
  type = "console",
  miniframe = TRUE,
  outpath = "."
)
}
\arguments{
\item{name}{Name of experiment. (}

\item{secname}{Name of section.}

\item{level}{Latex markup of the document element.
Default: "section". Latex supports:
"section", "subsection", "subsubsection".
Used for indentation of output.}

\item{clearpage}{Boolean. Default: TRUE. 
Ignored.}

\item{type}{Output to console (type="console") 
or to latex files (type="latex" | "beamer")?  
Default: "console".}

\item{miniframe}{Boolean. Default: TRUE.
If false, turns of miniframes for this section.}

\item{outpath}{Path for latex output files. Default: "".}
}
\value{
0 (invisible).
}
\description{
Sections of an experiment
}
\examples{
experimentSection(name="Experiment A", 
                  secname="Design", level="section", clearpage=FALSE)
                 
}
\seealso{
Other Report of Experiment: 
\code{\link{experimentAllXegaParameters}()},
\code{\link{experimentAnalysisOfTreatment}()},
\code{\link{experimentAppendix}()},
\code{\link{experimentBoxPlot}()},
\code{\link{experimentDesign}()},
\code{\link{experimentEffectReport}()},
\code{\link{experimentEnd}()},
\code{\link{experimentGrammarTable}()},
\code{\link{experimentHypothesisTest}()},
\code{\link{experimentIsNormal}()},
\code{\link{experimentMeanMatrix}()},
\code{\link{experimentNNTables}()},
\code{\link{experimentPlotPopStats}()},
\code{\link{experimentSolutionGenotype}()},
\code{\link{experimentSolutionTable}()},
\code{\link{experimentStart}()},
\code{\link{experimentStatistic}()},
\code{\link{experimentText}()},
\code{\link{experimentTreatmentParameters}()},
\code{\link{experimentTreatments}()},
\code{\link{experimentXegaParameters}()}
}
\concept{Report of Experiment}
