% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentElements.R
\name{experimentStart}
\alias{experimentStart}
\title{Start of experiment.}
\usage{
experimentStart(
  name = "Missing",
  title = "Undefined",
  author = "Unknown",
  purpose = "Unkown",
  beamertheme = "Berlin",
  beamercolor = "electricultramarine",
  type = "console",
  outpath = "."
)
}
\arguments{
\item{name}{Name of experiment. Default: "Missing".}

\item{title}{Title of the experiment. Default: "Undefined".}

\item{author}{Author of document. Default: "Unknown".}

\item{purpose}{Purpose of experiment. Default: "Unknown".}

\item{beamertheme}{Beamer layout. Default: "Berlin".
Choose from "default", "AnnArbor", "Antibes", "Bergen", 
"Berkeley", "Berlin", "Boadilla", "CambridgeUS", 
"Copenhagen", "Darmstadt", "Dresden", "Frankfurt", 
"Goettingen", "Hannover", "Ilmenau", "JuanLesPins", 
"Luebeck", "Madrid", "Malmoe", "Marburg", "Montpellier",
"PaloAlto", "Pittsburgh", "Rochester", "Singapore", 
"Szeged", "Warsaw".}

\item{beamercolor}{Color of beamer structure elements. 
            Setting \code{beamercolor="default"} uses the 
            standard color scheme of the beamer theme.
  Default: "electricultramarine".
\itemize{
\item Yellow(s): "goldenyellow", "electricyellow", "icterine",
                 "flavescent", "lemon".
\item Orange(s): "amber", "cadmiumorange", "internationalorange".
\item Red(s): "ferrarired", "fireenginered", "cadmiumred".
\item Blue(s): "ao", "babyblueeyes", "bleudefrance", "blue",
               "cobalt", "darkmidnightblue", "brandeisblue",
               "deepskyblue", "iris", "navyblue", "ultramarine",
               "electricultramarine".
\item Green(s): "darkpastelgreen", "cadmiumgreen".
 }}

\item{type}{Output to console (type="console") 
or to latex files (type="latex" | "beamer")?  
Default: "console".}

\item{outpath}{Path for latex output files. Default: "".}
}
\value{
0 (invisible).
}
\description{
Start of experiment.
}
\examples{
experimentStart(name="Experiment A.", title="First test", author="U.N. Owen")

}
\seealso{
Other Report of Experiment: 
\code{\link{experimentAllXegaParameters}()},
\code{\link{experimentAnalysisOfTreatment}()},
\code{\link{experimentAppendix}()},
\code{\link{experimentBoxPlot}()},
\code{\link{experimentDesign}()},
\code{\link{experimentEffectReport}()},
\code{\link{experimentEnd}()},
\code{\link{experimentGrammarTable}()},
\code{\link{experimentHypothesisTest}()},
\code{\link{experimentIsNormal}()},
\code{\link{experimentMeanMatrix}()},
\code{\link{experimentNNTables}()},
\code{\link{experimentPlotPopStats}()},
\code{\link{experimentSection}()},
\code{\link{experimentSolutionGenotype}()},
\code{\link{experimentSolutionTable}()},
\code{\link{experimentStatistic}()},
\code{\link{experimentText}()},
\code{\link{experimentTreatmentParameters}()},
\code{\link{experimentTreatments}()},
\code{\link{experimentXegaParameters}()}
}
\concept{Report of Experiment}
