% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentElements.R
\name{experimentText}
\alias{experimentText}
\title{Description (a block of text).}
\usage{
experimentText(
  ...,
  header = "",
  name = "",
  block = TRUE,
  type = "console",
  outpath = "."
)
}
\arguments{
\item{...}{A list of comma separated strings.}

\item{header}{Name of the block.}

\item{name}{Default: "".}

\item{block}{Default: TRUE: frame(block(text)) else
FALSE: frame(text).}

\item{type}{Output to console (type="console") 
or to latex files (type="latex" | "beamer")?  
Default: "console".}

\item{outpath}{Path for latex output files. Default: "".}
}
\value{
0 (invisible).
}
\description{
Description (a block of text).
}
\examples{
experimentText(
"This text describes my experiment.", 
"However, it is really neat.") 
                 
}
\seealso{
Other Report of Experiment: 
\code{\link{experimentAllXegaParameters}()},
\code{\link{experimentAnalysisOfTreatment}()},
\code{\link{experimentAppendix}()},
\code{\link{experimentBoxPlot}()},
\code{\link{experimentDesign}()},
\code{\link{experimentEffectReport}()},
\code{\link{experimentEnd}()},
\code{\link{experimentGrammarTable}()},
\code{\link{experimentHypothesisTest}()},
\code{\link{experimentIsNormal}()},
\code{\link{experimentMeanMatrix}()},
\code{\link{experimentNNTables}()},
\code{\link{experimentPlotPopStats}()},
\code{\link{experimentSection}()},
\code{\link{experimentSolutionGenotype}()},
\code{\link{experimentSolutionTable}()},
\code{\link{experimentStart}()},
\code{\link{experimentStatistic}()},
\code{\link{experimentTreatmentParameters}()},
\code{\link{experimentTreatments}()},
\code{\link{experimentXegaParameters}()}
}
\concept{Report of Experiment}
