% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentElements.R
\name{experimentTreatmentParameters}
\alias{experimentTreatmentParameters}
\title{A table of the parameters and their values of a treatment.}
\usage{
experimentTreatmentParameters(
  treatmentName,
  inpath = ".",
  outpath = ".",
  caption = "",
  name = "",
  type = "console"
)
}
\arguments{
\item{treatmentName}{Name of treatment.}

\item{inpath}{Path to treatment rds-files of experiment.
Default: ".".}

\item{outpath}{Path to report of experiment. Default: "."}

\item{caption}{Caption of table. Default: "".
Default: ".".}

\item{name}{Default: "".}

\item{type}{Output to console (type="console") 
or to latex files (type="latex" | "beamer")?  
Default: "console".}
}
\value{
Invisible 0.
}
\description{
A table of the parameters and their values of a treatment.
}
\examples{
cat("TBD!\n")

}
\seealso{
Other Report of Experiment: 
\code{\link{experimentAllXegaParameters}()},
\code{\link{experimentAnalysisOfTreatment}()},
\code{\link{experimentAppendix}()},
\code{\link{experimentBoxPlot}()},
\code{\link{experimentDesign}()},
\code{\link{experimentEffectReport}()},
\code{\link{experimentEnd}()},
\code{\link{experimentGrammarTable}()},
\code{\link{experimentHypothesisTest}()},
\code{\link{experimentIsNormal}()},
\code{\link{experimentMeanMatrix}()},
\code{\link{experimentNNTables}()},
\code{\link{experimentPlotPopStats}()},
\code{\link{experimentSection}()},
\code{\link{experimentSolutionGenotype}()},
\code{\link{experimentSolutionTable}()},
\code{\link{experimentStart}()},
\code{\link{experimentStatistic}()},
\code{\link{experimentText}()},
\code{\link{experimentTreatments}()},
\code{\link{experimentXegaParameters}()}
}
\concept{Report of Experiment}
