% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileIO.R
\name{fullFileNames}
\alias{fullFileNames}
\title{A list of paths to treatment rds-files.}
\usage{
fullFileNames(treatments, inpath = ".")
}
\arguments{
\item{treatments}{List of treatment names of experiments.}

\item{inpath}{Path for treatment files. Default: ".".}
}
\value{
A list of paths to treatment rds-files.
}
\description{
A list of paths to treatment rds-files.
}
\examples{
cat("TBD!\n")

}
\seealso{
Other File I/O: 
\code{\link{newFileName}()},
\code{\link{writeText}()}
}
\concept{File I/O}
