% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentDesign.R
\name{getTreatment}
\alias{getTreatment}
\title{Get treatment.}
\usage{
getTreatment(treatmentname = "", inpath = ".")
}
\arguments{
\item{treatmentname}{Name of the rds-file of a treatment.}

\item{inpath}{Path for treatment files. Default: ".".}
}
\value{
A treatment.
}
\description{
Get treatment.
}
\examples{
cat("TBD!\n")

}
\seealso{
Other Design of Experiment: 
\code{\link{designOfAllTreatments}()},
\code{\link{designOfExperiment}()},
\code{\link{designOfTreatment}()},
\code{\link{designSplitParameter}()},
\code{\link{pl2df}()}
}
\concept{Design of Experiment}
