% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentWrapper.R
\name{iterationReporter}
\alias{iterationReporter}
\title{Reporter of mean treatment performance data (fitness, generations) after k trials.}
\usage{
iterationReporter(
  experimentName,
  treatmentName,
  trials,
  i,
  resultDF,
  verbose = 0
)
}
\arguments{
\item{experimentName}{Name of experiment.}

\item{treatmentName}{Name of treatement.}

\item{trials}{Number of trails.}

\item{i}{Number of iteration.}

\item{resultDF}{Data frame of results.}

\item{verbose}{Progress report. 
If not \code{0}, output performance data to console.
Default: 0 - Not output to console.}
}
\value{
Invisible zero.
}
\description{
If \code{!(verbose==0)} print data on console.
}
\seealso{
Other Wrapper (Internal): 
\code{\link{treatmentReporter}()}
}
\concept{Wrapper (Internal)}
