% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kSymmetryWrapper.R
\docType{data}
\name{kSymmetryGrammarPath}
\alias{kSymmetryGrammarPath}
\title{The path to the grammars of the experiment.}
\format{
An object of class \code{character} of length 1.
}
\usage{
kSymmetryGrammarPath
}
\description{
For convenience. Should not be on cran.
}
\keyword{datasets}
