% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexElements.R
\name{latexBoxPlot}
\alias{latexBoxPlot}
\title{Produces box-and-whisker plots of treatments of an experiment (eps-file).}
\usage{
latexBoxPlot(
  treatments,
  variable,
  title = "",
  outline = TRUE,
  notch = FALSE,
  horizontal = TRUE,
  varwidth = FALSE,
  name = "",
  inpath = ".",
  outpath = "."
)
}
\arguments{
\item{treatments}{List of filenames of treatments.}

\item{variable}{Name of variable.}

\item{title}{Title of Boxplot. Default: "".}

\item{outline}{Boolean. If FALSE, remove outliers. Default: TRUE.}

\item{notch}{Boolean. If TRUE, notches are drawn. Default: FALSE.
Nonoverlapping notches indicate different 
medians of treatments.}

\item{horizontal}{Boolean. Default: TRUE. Plots are horizontal.}

\item{varwidth}{Boolean. Default: FALSE.
If TRUE,  width of the plot indicates sample size.}

\item{name}{Name of experiment. Default: "".}

\item{inpath}{Path to treatment files of experiment.
Default: ".".}

\item{outpath}{Path to report files.}
}
\value{
Name of eps file with output.
}
\description{
Produces box-and-whisker plots of treatments of an experiment (eps-file).
}
\examples{
cat("Example: TODO.\n")

}
\seealso{
Other Latex Elements: 
\code{\link{beamerSection}()},
\code{\link{latexEndDocument}()},
\code{\link{latexFigure}()},
\code{\link{latexMetaTable}()},
\code{\link{latexProlog}()},
\code{\link{latexSection}()},
\code{\link{latexTable}()},
\code{\link{latexTables}()},
\code{\link{latexText}()}
}
\concept{Latex Elements}
