% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexElements.R
\name{latexEndDocument}
\alias{latexEndDocument}
\title{End of document (LaTeX)}
\usage{
latexEndDocument()
}
\value{
A text string with latex markup.
}
\description{
End of document (LaTeX)
}
\examples{
a<-latexEndDocument()
print(a)
cat(a)
}
\seealso{
Other Latex Elements: 
\code{\link{beamerSection}()},
\code{\link{latexBoxPlot}()},
\code{\link{latexFigure}()},
\code{\link{latexMetaTable}()},
\code{\link{latexProlog}()},
\code{\link{latexSection}()},
\code{\link{latexTable}()},
\code{\link{latexTables}()},
\code{\link{latexText}()}
}
\concept{Latex Elements}
