% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexElements.R
\name{latexFigure}
\alias{latexFigure}
\title{The figure environment for a latex document.}
\usage{
latexFigure(
  name,
  fnName,
  caption = "No caption defined",
  label = "",
  width = 0.6,
  angle = -90
)
}
\arguments{
\item{name}{Name of the experiment.}

\item{fnName}{File name of the eps file of the graphic.}

\item{caption}{Caption of the figure.}

\item{label}{Label of the figure.}

\item{width}{Proportion of textwidth. Default: 0.6.}

\item{angle}{Angle. Default: -90.}
}
\value{
A text string with latex markup.
}
\description{
The figure environment for a latex document.
}
\examples{
a<-latexFigure(name="Ex A", fnName="test.eps", caption="test", label="ltest",
               width=0.6, angle=-90)
print(a)
cat(a)
}
\seealso{
Other Latex Elements: 
\code{\link{beamerSection}()},
\code{\link{latexBoxPlot}()},
\code{\link{latexEndDocument}()},
\code{\link{latexMetaTable}()},
\code{\link{latexProlog}()},
\code{\link{latexSection}()},
\code{\link{latexTable}()},
\code{\link{latexTables}()},
\code{\link{latexText}()}
}
\concept{Latex Elements}
