% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexElements.R
\name{latexMetaTable}
\alias{latexMetaTable}
\title{The meta table environment for a latex document.}
\usage{
latexMetaTable(name, fnName, caption = "No caption defined", label = "")
}
\arguments{
\item{name}{Name of the experiment.}

\item{fnName}{File name of the exported table element.}

\item{caption}{Caption of the table.}

\item{label}{Label of the table.}
}
\value{
A text string with latex markup.
}
\description{
The meta table environment for a latex document.
}
\examples{
a<-latexMetaTable(name="Ex A", fnName="test.tex", caption="test", label="ltest")
print(a)
cat(a)

}
\seealso{
Other Latex Elements: 
\code{\link{beamerSection}()},
\code{\link{latexBoxPlot}()},
\code{\link{latexEndDocument}()},
\code{\link{latexFigure}()},
\code{\link{latexProlog}()},
\code{\link{latexSection}()},
\code{\link{latexTable}()},
\code{\link{latexTables}()},
\code{\link{latexText}()}
}
\concept{Latex Elements}
