% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexElements.R
\name{latexProlog}
\alias{latexProlog}
\title{Prolog of minimal latex document.}
\usage{
latexProlog(
  name = "Missing",
  title = "Undefined",
  author = "Unknown",
  purpose = "Unknown"
)
}
\arguments{
\item{name}{Name of experiment. Default: "Missing".}

\item{title}{Title of the experiment. Default: "Undefined".}

\item{author}{Author of document. Default: "Unknown".}

\item{purpose}{Purpose of experiment. A short description. 
Default: "Unknown"}
}
\value{
A text string with LaTex markup.
}
\description{
Contains title, author and purpose of experiment.
}
\examples{
a<-latexProlog(name="Experiment A.", 
               title="First test", author="U.N. Owen",
               purpose="To show the impossible.\n")
cat(a)
}
\seealso{
Other Latex Elements: 
\code{\link{beamerSection}()},
\code{\link{latexBoxPlot}()},
\code{\link{latexEndDocument}()},
\code{\link{latexFigure}()},
\code{\link{latexMetaTable}()},
\code{\link{latexSection}()},
\code{\link{latexTable}()},
\code{\link{latexTables}()},
\code{\link{latexText}()}
}
\concept{Latex Elements}
