% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexElements.R
\name{latexSection}
\alias{latexSection}
\title{Sections of a latex document.}
\usage{
latexSection(name, level = "section", clearpage = TRUE)
}
\arguments{
\item{name}{Name of the document element.}

\item{level}{Latex markup of the document element.
Default: "section". Latex supports:
"section", "subsection", "subsubsection".}

\item{clearpage}{Boolean. Default: TRUE. 
Clearpage before start of document element.
Forces output of all floating elements of 
the previous document element.}
}
\value{
A text string with latex markup.
}
\description{
Sections of a latex document.
}
\examples{
latexSection(name="Design", level="section", clearpage=FALSE)
a<-latexSection(name="Design", level="section", clearpage=TRUE)
print(a)
cat(a)
}
\seealso{
Other Latex Elements: 
\code{\link{beamerSection}()},
\code{\link{latexBoxPlot}()},
\code{\link{latexEndDocument}()},
\code{\link{latexFigure}()},
\code{\link{latexMetaTable}()},
\code{\link{latexProlog}()},
\code{\link{latexTable}()},
\code{\link{latexTables}()},
\code{\link{latexText}()}
}
\concept{Latex Elements}
