% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexElements.R
\name{latexTable}
\alias{latexTable}
\title{The table environment for a latex document.}
\usage{
latexTable(
  df,
  caption = "",
  label = "",
  align = "",
  center = "center",
  sanitize = FALSE,
  tableLength = 60
)
}
\arguments{
\item{df}{Dataframe or matrix.}

\item{caption}{Caption of the table. Default: "".}

\item{label}{Label of the table.   Default: "".}

\item{align}{Alignment of columns. Default: "".}

\item{center}{Center table. Default: "center".}

\item{sanitize}{Boolean. Default: FALSE.}

\item{tableLength}{Integer. Default: 60.}
}
\value{
A text string with latex markup.
}
\description{
The table environment for a latex document.
}
\examples{
df<-data.frame(Name=c("Alice", "Bob"), Age=c(25, 27))
a<-latexTable(df, caption="Alice and Bob", label="AB")
print(a)

}
\seealso{
Other Latex Elements: 
\code{\link{beamerSection}()},
\code{\link{latexBoxPlot}()},
\code{\link{latexEndDocument}()},
\code{\link{latexFigure}()},
\code{\link{latexMetaTable}()},
\code{\link{latexProlog}()},
\code{\link{latexSection}()},
\code{\link{latexTables}()},
\code{\link{latexText}()}
}
\concept{Latex Elements}
