% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentElements.R
\name{latexTableLength}
\alias{latexTableLength}
\title{latex table length.}
\usage{
latexTableLength()
}
\value{
Integer. Number of lines in beamer table.
}
\description{
latex table length.
}
\examples{
latexTableLength()

}
\seealso{
Other Internal Configuration: 
\code{\link{beamerTableLength}()}
}
\concept{Internal Configuration}
