% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexElements.R
\name{latexTables}
\alias{latexTables}
\title{Write table(s) and meta-table(s) for a latex document.}
\usage{
latexTables(
  df,
  name,
  fnName,
  mfnName = "main",
  caption = "",
  label = "",
  align = "",
  center = "center",
  sanitize = FALSE,
  tableLength = 60,
  path = ""
)
}
\arguments{
\item{df}{Dataframe.}

\item{name}{Name of the experiment.}

\item{fnName}{File name of the exported table element.}

\item{mfnName}{File name of the exported meta table element.
Default: "main".}

\item{caption}{Caption of the table. Default: "".}

\item{label}{Label of the table.   Default: "".}

\item{align}{Alignment of columns. Default: "".}

\item{center}{Center table. Default: "center".}

\item{sanitize}{Boolean. Default: FALSE.}

\item{tableLength}{Integer. Default: 60.}

\item{path}{Path. Default: outpath.}
}
\value{
Invisible 0.
}
\description{
Write table(s) and meta-table(s) for a latex document.
}
\examples{
df<-data.frame(Name=c("Alice", "Bob"), Age=c(25, 27))
a<-latexTable(df, caption="Alice and Bob", label="AB")
print(a)

}
\seealso{
Other Latex Elements: 
\code{\link{beamerSection}()},
\code{\link{latexBoxPlot}()},
\code{\link{latexEndDocument}()},
\code{\link{latexFigure}()},
\code{\link{latexMetaTable}()},
\code{\link{latexProlog}()},
\code{\link{latexSection}()},
\code{\link{latexTable}()},
\code{\link{latexText}()}
}
\concept{Latex Elements}
