% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexElements.R
\name{latexText}
\alias{latexText}
\title{Text.}
\usage{
latexText(..., header = "")
}
\arguments{
\item{...}{A list of text strings.}

\item{header}{Name of the text.}
}
\value{
A text string.
}
\description{
Text.
}
\examples{
a<-latexText("Here", "we",  "are")
print(a)
cat(a)
}
\seealso{
Other Latex Elements: 
\code{\link{beamerSection}()},
\code{\link{latexBoxPlot}()},
\code{\link{latexEndDocument}()},
\code{\link{latexFigure}()},
\code{\link{latexMetaTable}()},
\code{\link{latexProlog}()},
\code{\link{latexSection}()},
\code{\link{latexTable}()},
\code{\link{latexTables}()}
}
\concept{Latex Elements}
