% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeExperiments.R
\name{mergeTreatments}
\alias{mergeTreatments}
\title{Merge treatment files of an experiment.}
\usage{
mergeTreatments(inpath = ".", outpath = ".")
}
\arguments{
\item{inpath}{Path of treatment files.}

\item{outpath}{Path for merged files.}
}
\value{
Invisible zero.
}
\description{
\code{inpath} is supposed to contain all rds-files of all treatments of an experiment.
             The filenames of a treatment are supposed to have the format "<treatmentname><serial>.rds".
             Serial is a number with three decimal places.
             \code{mergeTreatments()} reads all rds-files of a treatment, builds a new dataframe, 
             and writes the result to a rds-file with the name "merge<treatmentname>.rds" in 
             \code{outpath}.
}
\examples{
## dir.create(tempfile())
## dir.create(tempfile())
## tmpPath<-tempdir()
## subdirs<-list.files(path=tmpPath)
## dataTmpPath<-paste0(tmpPath, .Platform$file.sep, subdirs[1])
## mergeTmpPath<-paste0(tmpPath, .Platform$file.sep, subdirs[2])
tmpPath<-tempdir()
dataTmpPath<-tmpPath
mergeTmpPath<-tmpPath
unlink(paste0(tmpPath, "/*"))
list.files(path=dataTmpPath)
sym4replay19<-kSym(treatmentname="BoolT1SGPreplayXIX",
              experimentname="Ex", k=2, trials=3,
              grammarfn="AndOrNotTuned1.txt", algorithm="sgp",
              popsize=20, generations=10,
              executionModel="Sequential", verbose=0,
              semantics="byValue", replay=19,
              Gpath=kSymmetryGrammarPath, outpath=dataTmpPath, everyk=1)
sym4replay0<-kSym(treatmentname="BoolT1SGPreplayZ",
              experimentname="Ex", k=2, trials=3,
              grammarfn="AndOrNotTuned1.txt", algorithm="sgp",
              popsize=20, generations=10,
              executionModel="Sequential", verbose=0,
              semantics="byValue", replay=0,
              Gpath=kSymmetryGrammarPath, outpath=dataTmpPath, everyk=1)
list.files(path=dataTmpPath)
mergeTreatments(inpath=dataTmpPath, outpath=mergeTmpPath) 
list.files(path=dataTmpPath)
list.files(path=mergeTmpPath)

}
\concept{Merge of Treatments}
