% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envKsymmetry.R
\name{newEnvKsymmetry}
\alias{newEnvKsymmetry}
\title{Constructor for a k-symmetry problem environment.}
\usage{
newEnvKsymmetry(k = 5)
}
\arguments{
\item{k}{Integer. Number of bits.}
}
\value{
A problem environment for the k-symmetry problem.
        A closure with the following elements:
        \enumerate{
        \item $name() The name of the problem environment.
        \item $k()    The number of bit of the problem.     
        \item $BuildTest(expr)  Returns the fitness function 
                       for the boolean expression \code{expr}.
        \item $dec2bin(x) Converts an integer into a boolean vector.
        \item $kSymmetry(v) Tests if the boolean vector \code{v} is 
                       symmetric. Returns \code{0} or \code{1}.   
        \item $f(expr, gene, lF) The fitness function of the 
                       k-symmetry problem. Tests \code{2^k} boolean 
                       vectors for the expression \code{expr}. 
                       Returns the number of test errors.
        \item $globalOptimum() The minimal number of errors is \code{0}.
        }
}
\description{
Constructor for a k-symmetry problem environment.
}
\examples{
env3symmetry<-newEnvKsymmetry(k=3)
env3symmetry$f("OR(AND(D1, D3), NOT(OR(D1, D3)))")
env3symmetry$f("OR(AND(D1, D3), AND(NOT(D1), NOT(D3)))")
env3symmetry$f("AND(AND(NOT(D1), NOT(D3)), AND(NOT(D2), NOT(D1)))")

}
\seealso{
Other Problem Environment Factory: 
\code{\link{newEnvKsymmetryNN}()}
}
\concept{Problem Environment Factory}
