% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envKsymmetryNN.R
\name{newEnvKsymmetryNN}
\alias{newEnvKsymmetryNN}
\title{Constructor for a k-symmetry feedforward NN problem environment.}
\usage{
newEnvKsymmetryNN(k = 5, topology = c(5, 20, 1))
}
\arguments{
\item{k}{Integer. Number of bits.}

\item{topology}{Vector of integers. 
\code{topology[1]} defines the number of neurons 
on the input layer. 
\code{topology[length(topology)]} defines the 
number of neurons on the output layer.}
}
\value{
A problem environment for the k-symmetry problem.
        A closure with at least (TODO) the following elements:
        \enumerate{
        \item $name() The name of the problem environment.
        \item $k()    The number of bit of the problem.     
        \item $BuildTest(expr)  Returns the fitness function 
                       for the boolean expression \code{expr}.
        \item $dec2bin(x) Converts an integer into a boolean vector.
        \item $kSymmetry(v) Tests if the boolean vector \code{v} is 
                       symmetric. Returns \code{0} or \code{1}.   
        \item $f(expr, gene, lF) The fitness function of the 
                       k-symmetry problem. Tests \code{2^k} boolean 
                       vectors for the expression \code{expr}. 
                       Returns the number of test errors.
        \item $globalOptimum() The minimal number of errors is \code{0}.
        }
}
\description{
Constructor for a k-symmetry feedforward NN problem environment.
}
\examples{
e3NN<-newEnvKsymmetryNN(k=3, topology=c(3, 9, 1))
e3NN$f(rndParms(e3NN$nNNparms()))

}
\seealso{
Other Problem Environment Factory: 
\code{\link{newEnvKsymmetry}()}
}
\concept{Problem Environment Factory}
