% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileIO.R
\name{newFileName}
\alias{newFileName}
\title{Generate a new filename with a version number.}
\usage{
newFileName(fn, ftype = "tex", expname = "", path = ".")
}
\arguments{
\item{fn}{A filename.}

\item{ftype}{A filetype. Default: "tex".}

\item{expname}{Name of an Experiment. Default: "".}

\item{path}{A filepath. Default: "".}
}
\value{
A filename
}
\description{
The filename is built by catenating the following elements:
    \enumerate{
    \item \code{path} The path of the file.
    \item \code{expname} is the name of the experiment. May be empty.
    \item \code{fn} is the filename.
    \item \code{i}  is a number. Is formatted to the 3 positions 
                    (with leading \code{0}s). 
    \item \code{type} indicates the filetype (e.g. tex or rds). 
    } 
    If a filename exits, \code{i} is incremented by \code{1}.
    \strong{Warning:} The operation is not atomic.
}
\examples{
path<-tempdir()
a<-newFileName(fn="test", ftype="rds", path=path)
saveRDS(a, file=a)
cat("Filename 1:",a, "\n")
b<-newFileName(fn="test", ftype="rds", path=path)
cat("Filename 2:",b, "\n")
saveRDS(b, file=b)
c<-newFileName(fn="test", ftype="rds", path=path)
cat("Filename 3:",c, "\n")

}
\seealso{
Other File I/O: 
\code{\link{fullFileNames}()},
\code{\link{writeText}()}
}
\concept{File I/O}
