% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoofWrapper.R
\name{newSmoofWrapper}
\alias{newSmoofWrapper}
\title{Factory for wrapping smoof functions.}
\usage{
newSmoofWrapper(smoofFN)
}
\arguments{
\item{smoofFN}{smoof function (usually a call to the generator of a 
smoof function.}
}
\value{
A list of functions (
        \code{$name()},
        \code{$bitlength()},
        \code{$genelength()},
        \code{$lb()},
        \code{$ub()},
        \code{$f(parm, gene=0, lF=0)}, ...)
        For details, see the interface description
        in the package description.
}
\description{
A minimal wrapper for smoof functions. 
         This is an R-example for the facade architectural pattern.
}
\details{
The R-package smoof provides a large number 
         of benchmark and test functions for single and 
         multiobjective optimization. See
        https://cran.r-project.org/web/packages/smoof/index.html
}
\examples{
require(smoof)
t1<-newSmoofWrapper(makeSchwefelFunction(10))
}
\references{
Bossek, Jakob (2017)
         smoof: Single- and Multi-Objective Optimization Test Functions.
          The R Journal, 9(1), 103-113.
}
\concept{Constructor for Problem Environments}
