% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNnFF.R
\name{NN}
\alias{NN}
\title{A feedforward neural network with topology \code{top}.}
\usage{
NN(parm, topology, AH = ReLU, AO = ReLU, data)
}
\arguments{
\item{parm}{A parameter vector of length \code{numberOfNNParms(top)}.}

\item{topology}{The topology of the neural network.}

\item{AH}{Activation function hidden layers. Default: ReLU.}

\item{AO}{Activation function output layer. Default: ReLU.}

\item{data}{An input data matrix. The column number must match \code{top[1]}.  E.g. for xor.}
}
\value{
A vector of activation values of the output layer.
}
\description{
A feedforward neural network with topology \code{top}.
}
\examples{
NN(rndParms(numberOfNNParms(c(2, 3, 1))), c(2, 3, 1), data=xorData)

}
\seealso{
Other Feedforward Neural Network: 
\code{\link{P2NN}()},
\code{\link{ReLU}()},
\code{\link{RsquareNN}()},
\code{\link{numberOfNNParms}()},
\code{\link{printNNweights}()},
\code{\link{rndParms}()}
}
\concept{Feedforward Neural Network}
