% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNnFF.R
\name{P2NN}
\alias{P2NN}
\title{Convert the parameter vector of a NN into a list of weight matrices.}
\usage{
P2NN(parm, topology)
}
\arguments{
\item{parm}{A parameter vector of 
length \code{numberOfNNParms(topology)}.}

\item{topology}{The topology of the neural network.}
}
\value{
A list of weight matrices.
}
\description{
The dimensions of the weight matrices are defined by 
             the vector \code{topology}.
}
\examples{
topology<-c(1, 3, 2, 1)
parms<-rndParms(numberOfNNParms(topology))
P2NN(parms, topology)

}
\seealso{
Other Feedforward Neural Network: 
\code{\link{NN}()},
\code{\link{ReLU}()},
\code{\link{RsquareNN}()},
\code{\link{numberOfNNParms}()},
\code{\link{printNNweights}()},
\code{\link{rndParms}()}
}
\concept{Feedforward Neural Network}
