% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consoleElements.R
\name{pause}
\alias{pause}
\title{Wait for input.}
\usage{
pause()
}
\value{
Keyboard input of user (invisible).
}
\description{
Wait for input.
}
\examples{
pause()
a<-pause()
print(a)

}
\seealso{
Other Console Elements: 
\code{\link{beamerText}()},
\code{\link{consoleBoxPlot}()},
\code{\link{consoleEndDocument}()},
\code{\link{consoleFigure}()},
\code{\link{consoleProlog}()},
\code{\link{consoleSection}()},
\code{\link{consoleTable}()},
\code{\link{consoleText}()}
}
\concept{Console Elements}
