% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentDesign.R
\name{pl2df}
\alias{pl2df}
\title{Convert a named parameter list into a dataframe.}
\usage{
pl2df(pl)
}
\arguments{
\item{pl}{Parameter list.}
}
\value{
A dataframe with one column: The parameter values.
        The rownames are the (raw) parameter names.
}
\description{
Three parameters need special handling:
 \enumerate{
    \item \code{$tRNG}: Convert to one string.
    \item \code{$penv}: Replace penv by name of penv.
    \item \code{$grammar}: Replace grammar by its name.
    }
}
\examples{
cat("TBD!\n")

}
\seealso{
Other Design of Experiment: 
\code{\link{designOfAllTreatments}()},
\code{\link{designOfExperiment}()},
\code{\link{designOfTreatment}()},
\code{\link{designSplitParameter}()},
\code{\link{getTreatment}()}
}
\concept{Design of Experiment}
