% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNnFF.R
\name{printNNweights}
\alias{printNNweights}
\title{Print the parameter vector of a NN with topology c(2, 3, 1)}
\usage{
printNNweights(parm, topology)
}
\arguments{
\item{parm}{A parameter vector.}

\item{topology}{A topology vector.}
}
\value{
Invisible zero.
}
\description{
TODO: Add row and column names.
}
\examples{
printNNweights(parm=rndParms(numberOfNNParms(c(2, 3, 1))), topology=c(2, 3, 1))

}
\seealso{
Other Feedforward Neural Network: 
\code{\link{NN}()},
\code{\link{P2NN}()},
\code{\link{ReLU}()},
\code{\link{RsquareNN}()},
\code{\link{numberOfNNParms}()},
\code{\link{rndParms}()}
}
\concept{Feedforward Neural Network}
