% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testsOfHypothesis.R
\name{rmOutliers}
\alias{rmOutliers}
\title{Removes outliers of data vector.}
\usage{
rmOutliers(x, coef = 1.5)
}
\arguments{
\item{x}{A numeric vector.}

\item{coef}{Scaling of threshold for outliers. Default: 1.5.}
}
\value{
A numeric vector.
}
\description{
An outlier is a datapoint outside \code{coef} times
             of the box size. See \code{boxplot.stats}.
}
\examples{
x<-rnorm(30, mean=0, sd=2)
rmOutliers(x, coef=0)
rmOutliers(x, coef=0.5)
rmOutliers(x, coef=1.0)
rmOutliers(x)

}
\concept{Transformation}
