% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNnFF.R
\name{rndParms}
\alias{rndParms}
\title{Generates a random vector of length n in an n-dimensional hypercube.}
\usage{
rndParms(n, lb = -1, ub = 1)
}
\arguments{
\item{n}{Number of elements.}

\item{lb}{Lower bound.}

\item{ub}{Upper bound.}
}
\value{
A random vector of length n with elements in [lb, ub].
}
\description{
Generates a random vector of length n in an n-dimensional hypercube.
}
\examples{
rndParms(numberOfNNParms(c(2, 3, 1)))

}
\seealso{
Other Feedforward Neural Network: 
\code{\link{NN}()},
\code{\link{P2NN}()},
\code{\link{ReLU}()},
\code{\link{RsquareNN}()},
\code{\link{numberOfNNParms}()},
\code{\link{printNNweights}()}
}
\concept{Feedforward Neural Network}
