% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNnFF.R
\name{RsquareNN}
\alias{RsquareNN}
\title{R-square.}
\usage{
RsquareNN(predicted, actual)
}
\arguments{
\item{predicted}{Predicted output vector.}

\item{actual}{Actual output vector.}
}
\value{
The sum of squared errors.
}
\description{
R-square.
}
\examples{
RsquareNN(c(1, 2, 3), c(0, 3, 3))

}
\seealso{
Other Feedforward Neural Network: 
\code{\link{NN}()},
\code{\link{P2NN}()},
\code{\link{ReLU}()},
\code{\link{numberOfNNParms}()},
\code{\link{printNNweights}()},
\code{\link{rndParms}()}
}
\concept{Feedforward Neural Network}
