% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentStatistics.R
\name{selectFromExperiment}
\alias{selectFromExperiment}
\title{Select a subset of the descriptive statistics of an experiment.}
\usage{
selectFromExperiment(
  treatments,
  treatment = NULL,
  variable = NULL,
  inpath = "."
)
}
\arguments{
\item{treatments}{List of treatment names of experiment.}

\item{treatment}{Name of treatment. Default: NULL.}

\item{variable}{Name of variable.  Defautl: NULL.}

\item{inpath}{Path to treatement rds-files of experiment.
Default: ".".}
}
\value{
A dataframe with the following columns:
        \enumerate{
        \item \code{$Treatment}: Name of treatment.
        \item \code{$Trials}:    Number of trials.
        \item \code{$Variable}:  Name of Variable.
        \item \code{$min}:       Minimum.
        \item \code{$mean}:      Mean.
        \item \code{$sd}:        Standard deviation. 
        \item \code{$max}:       Minimum.
                  }
}
\description{
Specify either treatment name or variable name or 
             none. Returns the appropriate subset of records of 
             the dataframe of all descriptive statistics of the 
             experiment.
}
\examples{
cat("TBD! \n")

}
\seealso{
Other Statistics of Experiment: 
\code{\link{summaryExperiment}()}
}
\concept{Statistics of Experiment}
