% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentStatistics.R
\name{summaryExperiment}
\alias{summaryExperiment}
\title{Dataframe of descriptive statistics for experiment.}
\usage{
summaryExperiment(treatments, byVars = TRUE, inpath = ".")
}
\arguments{
\item{treatments}{List of treatment names of experiment.}

\item{byVars}{Boolean. If TRUE, Sort by variables. 
Default: TRUE.}

\item{inpath}{Path to treatement rds-files of experiment.
Default: ".".}
}
\value{
A dataframe with the following columns:
        \enumerate{
        \item \code{$Treatment}: Name of treatment.
        \item \code{$Trials}:    Number of trials.
        \item \code{$Variable}:  Name of Variable.
        \item \code{$min}:       Minimum.
        \item \code{$mean}:      Mean.
        \item \code{$sd}:        Standard deviation. 
        \item \code{$max}:       Minimum.
                  }
}
\description{
Dataframe of descriptive statistics for experiment.
}
\examples{
cat("TBD! \n")

}
\seealso{
Other Statistics of Experiment: 
\code{\link{selectFromExperiment}()}
}
\concept{Statistics of Experiment}
