% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentStatistics.R
\name{summaryTreatment}
\alias{summaryTreatment}
\title{Dataframe of descriptive statistics of all variables of a treatment.}
\usage{
summaryTreatment(name, df)
}
\arguments{
\item{name}{Name of treatment.}

\item{df}{Experimental results of treatment.
Each record is the result of one trial.}
}
\value{
A dataframe with the following columns:
        \enumerate{
        \item \code{$Treatment}: Name of treatment.
        \item \code{$Trials}:    Number of trials.
        \item \code{$Variable}:  Name of Variable.
        \item \code{$min}:       Minimum.
        \item \code{$mean}:      Mean.
        \item \code{$sd}:        Standard deviation. 
        \item \code{$max}:       Minimum.
                  }
}
\description{
Dataframe of descriptive statistics of all variables of a treatment.
}
\examples{
df<-data.frame(Fitness=sample(100, 10), Seconds=sample(50, 10), 
               Generations=sample(30, 10), Evaluations=sample(3000, 10)) 
summaryTreatment("MC-GP", df)

}
\seealso{
Other Descriptive Statistics (Internal): 
\code{\link{summaryVariable}()}
}
\concept{Descriptive Statistics (Internal)}
