% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentStatistics.R
\name{summaryVariable}
\alias{summaryVariable}
\title{Summarize a performance variable of one treatment of an experiment.}
\usage{
summaryVariable(name, variable, column)
}
\arguments{
\item{name}{Name of treatment.}

\item{variable}{Name of variable.}

\item{column}{Vector of values of variable.}
}
\value{
A dataframe with the following columns:
        \enumerate{
        \item \code{$Treatment}: Name of treatment.
        \item \code{$Trials}:    Number of trials.
        \item \code{$Variable}:  Name of Variable.
        \item \code{$min}:       Minimum.
        \item \code{$mean}:      Mean.
        \item \code{$sd}:        Standard deviation. 
        \item \code{$max}:       Minimum.
                  }
}
\description{
Summarize a performance variable of one treatment of an experiment.
}
\examples{
summaryVariable("MC-GP", "Seconds", sample(100, 10))

}
\seealso{
Other Descriptive Statistics (Internal): 
\code{\link{summaryTreatment}()}
}
\concept{Descriptive Statistics (Internal)}
