% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammarTuning.R
\name{sympairs}
\alias{sympairs}
\title{Returns a the set of production rules with symmetric variable pairs.}
\usage{
sympairs(varNT, varSym, k)
}
\arguments{
\item{varNT}{The non-terminal symbol on the left-hand side of a BNF rule.}

\item{varSym}{The symbol of the variable name. E.g. "V".}

\item{k}{The number of bits.}
}
\value{
A text string with the production rules in BNF.
}
\description{
Symmetric variable pairs are the pairs of variables 
             with the same truth value in the k-symmetry problem.
             E.g. In a 4-symmetry problem with variables V1-V4, 
             the 2 symmetric variable pairs are (V1, V4) and (V2, V3).
}
\details{
For the BNF notation, see the R-package \code{xegaBNF}
             The function is called by the preprocessor 
             \code{xegaBNF::preBNF} and is expected to return
             a part of a BNF grammar specification. 
             Used in the grammar AndOrNotTuned2.txt.
}
\examples{
sympairs(varNT="<sympairs>", varSym="V", k=4) 

}
\seealso{
Other Grammar Tuning: 
\code{\link{nandpairs}()},
\code{\link{sym2pairs}()}
}
\concept{Grammar Tuning}
