% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envKsymmetry.R
\docType{data}
\name{t1}
\alias{t1}
\title{A test case for the 5-symmetry problem.}
\format{
An object of class \code{character} of length 1.
}
\usage{
t1
}
\description{
A test case for the 5-symmetry problem.
}
\keyword{datasets}
