% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentWrapper.R
\name{treatmentReporter}
\alias{treatmentReporter}
\title{Reporter of and observer treatment performance and solution data.}
\usage{
treatmentReporter(experimentName, treatmentName, s, i, verbose = 0)
}
\arguments{
\item{experimentName}{Name of experiment.}

\item{treatmentName}{Name of treatment.}

\item{s}{A solution object of \code{xegaRun} (with elements
\code{$popStat}, \code{$fit}, \code{$solution}, 
\code{$evalFail}, \code{$GAconfig}, \code{$GAenv}, and \code{$timer}.}

\item{i}{Number of iteration.}

\item{verbose}{If not \code{0}, output treatment data to console.
Progress report.
Default: 0 - Not output to console.}
}
\value{
A dataframe with a single row and the following columns:
        \itemize{
        \item \code{$Fit}: The phenotype value of the solution.
        \item \code{$Solution}: The solution (its phenotype).
        \item \code{$Seconds}: Execution time of \code{xegaRun()} in seconds.
        \item \code{$Generations}: Number of generations until termination.
        \item \code{$Evaluations}: Number of evaluations until termination.
        }
}
\description{
Returns a dataframe with the performance data and 
             the phenotype of the solution of a single treatment.
             If \code{!(verbose==0)} print data on console.
}
\seealso{
Other Wrapper (Internal): 
\code{\link{iterationReporter}()}
}
\concept{Wrapper (Internal)}
