% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentWrapper.R
\name{treatmentRun}
\alias{treatmentRun}
\title{Runs a treatment with xegaRun repeatedly \code{trials} times.}
\usage{
treatmentRun(
  experimentName = "e0",
  treatmentName = "t0",
  tReplay = 0,
  trials = 1,
  everyK = 10,
  outpath = ".",
  tVerbose = 0,
  ...
)
}
\arguments{
\item{experimentName}{Name of experiment. Default: "e0".}

\item{treatmentName}{Name of treatment.  Default: "t0".}

\item{tReplay}{Integer. Seed of Random Number.  Default: 0 (No seeding)}

\item{trials}{Default: 1.}

\item{everyK}{Save results of trials after \code{everyK} trials (Default: 10).}

\item{outpath}{Path for treatment results. Deafult: ".".}

\item{tVerbose}{Screen output. Default: 0 (no output).
\itemize{
\item \code{tVerbose=0}: No output. 
\item \code{tVerbose=1}: Mean performance after i trials. 
\item \code{tVerbose>1}: Mean performance after i trials 
                        and performance of i-th trial.
}}

\item{...}{Arguments for xegaRun of this treatment.}
}
\value{
A named list with the following elements
  \itemize{
  \item \code{$lastGAResult}: The return object of \code{xegaRun()}.
  \item \code{$resultDF}: A dataframe with the following columns:
  \itemize{
    \item \code{$Fit}: The best fitness value of a trial.
    \item \code{$Solution}: The best solution (the phenotype of the gene).
    \item \code{$Seconds}:  Execution time of the trial in seconds. 
    \item \code{$Generations}: The number of generations until termination.
    \item \code{$Evaluations}: The number of fitness evaluations.   
    }
  \item \code{$tArgs}: Arguments of a treatmentRun:
      \code{$tRNG},
      \code{$tReplay},
      \code{$treatmentName},
      \code{$trials},
      \code{$everyK},
      \code{$outPath},
      \code{$batchPath},
      \code{$tVerbose}.
  \item \code{$xegaArgs}: Arguments of \code{xega::xegaRun}.
  \item \code{$filename}: The name of the last file written.
    }
}
\description{
Tentative. 
             Not yet flexible at all.
}
\details{
Repeats xegaRun several times. 
        \itemize{
         \item
         As a side effect writes files with the experimental data
         into the directory \code{outpath} which must exist.
         \item
         The treatmentname may not contain numbers, because
         the files generated \code{everyk} trials have serial 
         numbers with 3 decimal places at the end. 
         }
}
\examples{
require(xegaBNF)
tmpPath<-tempdir()
gBNF<-compileBNF(preBNF(booleanGrammarK, list(k=3)))
kSymEnv<-newEnvKsymmetry(k=3)
r<-treatmentRun(experimentName="ExpA", treatmentName="t0", tReplay=0, 
      trials=2, everyK=1, outpath=tmpPath, tVerbose=1,  
      penv=kSymEnv, grammar=gBNF, algorithm="sgp", max=FALSE, popsize=20, 
      generations= 5, crossrate=0.2, mutrate=0.4, 
      ivmutrate="Const", mutrate2=0.4,
      ivcrossrate="Const", crossrate2=0.2,
      executionModel="Sequential", verbose=1, 
      semantics="byValue")
print(r$resultDF)

}
\seealso{
Other Experiment: 
\code{\link{kSym}()},
\code{\link{kSymNN}()}
}
\concept{Experiment}
