% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kSymmetryWrapper.R
\name{variableNamesLHS}
\alias{variableNamesLHS}
\title{Generates a list of alternative variable names.}
\usage{
variableNamesLHS(varSym, k)
}
\arguments{
\item{varSym}{Variable symbol. E.g. "D"-}

\item{k}{Number of bits.}
}
\value{
Text.  Contains the LHS of a production rule with 
          a list of k variable names in EBNF.
}
\description{
Generates a list of alternative variable names.
}
\seealso{
Other xegaBNF: 
\code{\link{booleanGrammarK}()},
\code{\link{writeBNF}()}
}
\concept{xegaBNF}
