% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kSymmetryWrapper.R
\name{writeBNF}
\alias{writeBNF}
\title{Write a grammar file.}
\usage{
writeBNF(g, fn = NULL, eol = "\\n")
}
\arguments{
\item{g}{A grammar object.}

\item{fn}{A filename.}

\item{eol}{End of line symbol.}
}
\description{
Write a grammar file.
}
\seealso{
Other xegaBNF: 
\code{\link{booleanGrammarK}()},
\code{\link{variableNamesLHS}()}
}
\concept{xegaBNF}
