% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileIO.R
\name{writeText}
\alias{writeText}
\title{Write a text file.}
\usage{
writeText(text, fn = NULL)
}
\arguments{
\item{text}{A text string.}

\item{fn}{File name. Default: NULL.}
}
\value{
A text string (invisible)
}
\description{
Write a text file.
}
\examples{
path<-tempdir()
a<-newFileName(fn="test", ftype="txt", path=path)
b<-writeText(a, fn=a)

}
\seealso{
Other File I/O: 
\code{\link{fullFileNames}()},
\code{\link{newFileName}()}
}
\concept{File I/O}
