% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaPopDiag.R
\name{xegaPlotPopStats}
\alias{xegaPlotPopStats}
\title{Plot population fitness statistics}
\usage{
xegaPlotPopStats(
  xegaResult,
  epsfile = FALSE,
  filename = NULL,
  description = ""
)
}
\arguments{
\item{xegaResult}{the result of \code{RunSGA} (or \code{RunSGP}).}

\item{epsfile}{If \code{epsfile} is \code{FALSE}, then the 
 plot is shown on an X11-device else an epsfile
with the name of the problem environment is printed.}

\item{filename}{If specified, overrides filename of epsfile.}

\item{description}{Additional text.}
}
\value{
the return code of device.off()
}
\description{
\code{xegaPlotPopStats} plots the population 
             fitness statistics (min, Q1, mean, median, Q3, max)
             of a run of a genetic algorithm.

For an introduction to graphics in R, see 
Paul Murrell (2006) R-Graphics, Chapman & Hall/CRC, Boca Raton
}
\examples{
library(xega)
a<-xegaRun(Parabola2D)
xegaPlotPopStats(a)
\dontrun{xegaPlotPopStats(a, epsfile=TRUE)}

}
\references{
Murrell, Paul (2006):
      \emph{R-Graphics}. 
      Chapman & Hall/CRC, Boca Raton.

TODO: Automatic file names for printing several runs ...
}
\concept{genetic algorithm visualization}
\concept{visualization}
