% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNnFF.R
\docType{data}
\name{xorData}
\alias{xorData}
\title{Data for the xor problem.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 3 columns.
}
\usage{
xorData
}
\description{
Data for the xor problem.
}
\keyword{datasets}
